+++ backend/src/main/java/com/mapbefine/mapbefine/oauth/exception/OauthException.java
@@ -4,7 +4,7 @@ import com.mapbefine.mapbefine.common.exception.ErrorCode;
 import com.mapbefine.mapbefine.common.exception.NotFoundException;
 import com.mapbefine.mapbefine.oauth.domain.OauthServerType;
 
-public class OathException {
+public class OauthException {
 
     public static class OauthNotFoundException extends NotFoundException {
         public OauthNotFoundException(OauthErrorCode errorCode, OauthServerType oauthServerType) {
